Add job type
